third_party/WebKit/LayoutTests/fast/dom/dir-non-html-document.html

<!DOCTYPE html>
<html>
<head>
<link rel="help" href="http://www.whatwg.org/specs/web-apps/current-work/multipage/elements.html#dom-document-dir"/>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Make sure the dir attribute is exposed on non-HTML documents.");

var xmlDocument = document.implementation.createDocument("", "", null);
shouldBe("xmlDocument.__proto__", "XMLDocument.prototype");
shouldBe("xmlDocument.__proto__.__proto__", "Document.prototype");

shouldBeEqualToString("xmlDocument.dir", "");
shouldNotThrow("xmlDocument.dir = 'ltr'");
shouldBeEqualToString("xmlDocument.dir", "");
</script>
</body>
</html>
