third_party/blink/web_tests/fast/dom/dir-non-html-document-expected.txt

Make sure the dir attribute is exposed on non-HTML documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xmlDocument.__proto__ is XMLDocument.prototype
PASS xmlDocument.__proto__.__proto__ is Document.prototype
PASS xmlDocument.dir is ""
PASS xmlDocument.dir = 'ltr' did not throw exception.
PASS xmlDocument.dir is ""
PASS successfullyParsed is true

TEST COMPLETE
